+++ aide_design/designs/civil_work/floc_classes/bottom_baffle.py
@@ -1,14 +1,14 @@
-from aide_design.units import unit_registry as u
+from aide_design.shared.units import unit_registry as u
 from aide_render.builder_classes import DP, HP
 
 
-class TopBaffle:
-    """This is a top baffle class. It's called by the top baffles assembly
+class BottomBaffle:
+    """This is a bottom baffle class. It's called by the bottom baffles assembly
     class so that the hierarchy of objects in Python is the same as in Fusion.
 
     """
 
-    def __init__(self, L_top_baffle, baffle_thickness, W_chan):
+    def __init__(self, L_bottom_baffle, baffle_thickness, W_chan):
         """This is where the "instantiation" occurs. Think of this as "rendering the
         template" or "using the cookie-cutter to make the cookie". Here is where we
         call all the methods that determine design qualities of the specific
@@ -17,8 +17,8 @@ class TopBaffle:
         Parameters
         ----------
 
-        L_top_baffle : float
-            Length of the baffles at the top of the flocculator
+        L_bottom_baffle : float
+            Length of the baffles on the bottom of the flocculator
 
         baffle_thickness : float
             Thickness of a baffle
@@ -28,6 +28,6 @@ class TopBaffle:
 
         """
 
-        self.Height = DP(L_top_baffle.magnitude, L_top_baffle.units)
+        self.Height = DP(L_bottom_baffle.magnitude, L_bottom_baffle.units)
         self.Thickness = DP(baffle_thickness.magnitude, baffle_thickness.units)
         self.Width = DP(W_chan.magnitude, W_chan.units)
